requirements.txt
fastapi[all]
uvicorn
cassandra-driver